refactor the nested ncdatasets
